Update ofUtils_functions.markdown
Correction of (probably old?) function description:
`result=ofSplitString("of rocks", "of");` does not result in result[0]=r and result[1]=cks. (It would rather be [0]="" and [1]=" rocks")
Exchanged example with ofSplitString("of rocks", "ro"); for clarity of explanation.
Merged this corrected old description including the example with the existing, but shorter new description.