updated vs.md to prevent exception error
If you follow the install guide for Visual Studio users on the website, you'll run into an error when you want to create an OF project, because it didn't specify, that Common Tools for Visual C++ 2015 have to be installed. The problem is documented in the forums, the website just didn't reflect it yet. Hope this fixes it.