Removes hover effect from method names
This is something that has bothered me for the longest time.

The <h1> names of methods/variables/functions in the documentation have
a hover effect, just like the menu items. But these items are not
actually links! I always find this confusing when I want to copy the
name of a function, thinking I can't doubleclick to select the name and
have to drag-select instead (as clicking would open the non-existent
link).

This change removes the hover effect by adding the already existing
'nohover' css class. No other elements are changed.