Update to current Android Studio setup steps
The steps shown on this page have fallen behind what is needed currently. The instructions to "edit the file libs/openFrameworksCompiled/project/android/paths.make" are no longer applicable as paths.make doesn't even exist and that's not the current way to tell Android Studio where the NDK is. So I took the relevant updated lines from the INSTALL_ANDROID_STUDIO.md file in OpenFrameworks0.10.0 for Android and pasted them here in what seems the appropriate place, and removed the obsolete lines. 

I also think probably there are other obsolete details on this page, which could be copied from that more updated file, but the format was not identical and I didn't want to make a mistake, so I just changed what I was confident about. 

I notice that the seeming links to the NDK versions don't work as links - clicking them takes you back to the help page you are already on, so you have to copy the link text to download them from those. Not sure how/if you'd want that fixed exactly, so I didn't go making them links.