Make docker env for development, update nikola [WIP]
The build instructions in current README fail due to dependency issues on up-to-date MacOS, Ubuntu and Alpine Linux.

This is a WIP attempt to provide an easy-to-install (and cross-platform) docker environment for development work.

To get the dependencies to work, I had to upgrade nikola (which dragged some plugins along with it).

Dependencies are unpinned at the moment, until it's proven to be working correctly.

At first glance it seems to work, but before taking it any further I want feedback in case this is a wrong direction.